fix : fix layout order, quote, and attibutet

+++ src/features/pages/catalogue/attributes/index.tsx
@@ -17,19 +17,37 @@ import { useAtom } from 'jotai'
 import { deleteAttribute } from '@/services/attributes-service'
 import { useNotificationAntd } from '@/components/toast'
 import { attributeAtom } from '@/store/AttributeAtom'
+import { AddIcon, TrashIconRed, PencilIconBlue } from '@public/icon'
+import ButtonDelete from '@/components/button/ButtonAction'
+import Pagination from '@/components/pagination'
+import ButtonIcon from '@/components/button/ButtonIcon'
+import SearchTable from '@/components/search/SearchTable'
+import ShowPageSize from '@/components/pagination/ShowPageSize'
+import ModalDelete from '@/components/modal/ModalDelete'
+import { useRouter } from 'next/navigation'
+import Image from 'next/image'
 
 const index = ({ attributesData }: { attributesData?: any }) => {
+    const router = useRouter()
+
     const { contextHolder, notifySuccess } = useNotificationAntd()
     const [data, setData] = useAtom(attributeAtom)
     const [filteredData, setFilteredData] = useState<AttributesType[]>([])
     const [search, setSearch] = useState('')
+    const [currentPage, setCurrentPage] = useState(1);
+    const [pageSize, setPageSize] = useState(10);
+    const [isOpenModalFilter, setisOpenModalFilter] = useState(false)
+    const [selectedRowKeys, setSelectedRowKeys] = useState<React.Key[]>([]);
+    const [openModalDelete, setOpenModalDelete] = useState(false)
+    const [deletedData, setDeletedData] = useState<any>(null)
 
     const handleDelete = async (id: any) => {
+        if (!deletedData) return;
         try {
             const res = await deleteAttribute(id)
             if (res.success == true) {
                 notifySuccess(res.message)
-                const updatedData = data.filter(item => item.id !== id);
+                const updatedData = data.filter(item => item.id !== deletedData);
                 setData(updatedData);
 
                 setFilteredData(updatedData.filter(item =>
@@ -106,16 +124,32 @@ const index = ({ attributesData }: { attributesData?: any }) => {
                 );
 
                 return (
-                    <Dropdown overlay={menu} trigger={['click']} >
-                        <button className="flex items-center justify-center px-2 py-1 border rounded ">
-                            Actions <MoreOutlined className="ml-1" />
-                        </button>
-                    </Dropdown >
+                    <div className="flex items-center justify-end gap-3 pe-4">
+                        <ButtonIcon
+                            color='primary'
+                            variant='filled'
+                            size="small"
+                            icon={PencilIconBlue}
+                            onClick={() => router.push(routes.eCommerce.editBrands(row.id))}
+                        />
+                        <ButtonIcon
+                            color='danger'
+                            variant='filled'
+                            size="small"
+                            icon={TrashIconRed}
+                            onClick={() => handleOpenModalDelete(row.id)}
+                        />
+                    </div >
                 );
             }
         },
 
     ]
+
+    const handleOpenModalDelete = (data: any) => {
+        setOpenModalDelete(true)
+        setDeletedData(data)
+    }
 
     const handleSearch = (value: string) => {
         const search = value.toLowerCase()
@@ -137,33 +171,83 @@ const index = ({ attributesData }: { attributesData?: any }) => {
     }, [attributesData, search])
     return (
         <>
+            <ModalDelete
+                open={openModalDelete}
+                onCancel={() => setOpenModalDelete(false)}
+                onDelete={handleDelete}
+                item='attribute'
+            />
             {contextHolder}
             <div className="mt-6 mx-4 mb-0">
-                <h1 className='text-xl font-bold'>
-                    Attributes
-                </h1>
-                <Breadcrumb
-                    items={breadcrumb}
-                />
+                <div className='flex justify-between items-center'>
+                    <div>
+                        <h1 className='text-xl font-bold'>
+                            Attributes
+                        </h1>
+                        <Breadcrumb
+                            items={breadcrumb}
+                        />
+                    </div>
+                    <Button
+                        icon={<Image
+                            src={AddIcon}
+                            alt='add-icon'
+                            width={15}
+                            height={15}
+                        />}
+
+                        label='Add Attribute'
+                        link={routes.eCommerce.createAttributes}
+
+                    />
+                </div>
             </div>
             <Content className="mt-6 mx-4 mb-0">
                 <div style={{ padding: 24, minHeight: 360, background: '#fff' }}>
-                    <div className='flex justify-end mb-4'>
+
+                    <div className='flex justify-between mb-4'>
                         <div className='flex items-center gap-2'>
-                            <SearchInput onSearch={handleSearch} />
-                            <Button
-                                icon={<PlusCircleOutlined />}
-                                label='Add Attributes'
-                                link={routes.eCommerce.createAttributes}
+                            <ShowPageSize
+                                pageSize={pageSize}
+                                onChange={setPageSize}
+                            />
+                            <SearchTable
+                                value={search}
+                                onChange={(e) => setSearch(e.target.value)}
+                                onSearch={() => console.log('Searching for:', search)}
                             />
                         </div>
-
+                        {
+                            selectedRowKeys.length > 0 && <ButtonDelete
+                                label='Delete All'
+                                icon={<Image
+                                    src={TrashIconRed}
+                                    alt='trash-icon'
+                                    width={10}
+                                    height={10}
+                                />}
+                                onClick={() => setisOpenModalFilter(true)}
+                                position='start'
+                                style={{ padding: '1.2rem 1.7rem' }}
+                                btnClassname='btn-delete-all'
+                            />
+                        }
                     </div>
+
                     <Table
                         columns={columns}
                         dataSource={filteredData}
                         withSelectableRows
+                        selectedRowKeys={selectedRowKeys}
+                        onSelectChange={setSelectedRowKeys}
                     />
+                    <Pagination
+                        current={currentPage}
+                        total={filteredData.length}
+                        pageSize={pageSize}
+                        onChange={(page) => setCurrentPage(page)}
+                    />
+
                 </div>
             </Content>
         </>
